feat: Create UrlData entity to validate and store dynamic url data like query params and routes
