+++ src/app/services/request/entities/mod.rs
@@ -1,0 +1,3 @@
+pub mod methods;
+pub mod requests;
+pub mod url;
