+++ src/app/services/web_client/facade.rs
@@ -1,7 +1,7 @@
 use tokio::task::JoinHandle;
 
 use super::entities::Response;
-use crate::app::services::request::entities::RequestData;
+use crate::app::services::request::entities::requests::RequestData;
 
 pub trait WebClientFacade: Send {
     fn submit_async(&mut self, request: RequestData) -> JoinHandle<Result<Response, String>>;
