+++ src/app/services/web_client/service.rs
@@ -3,7 +3,8 @@ use tokio::task::JoinHandle;
 use super::entities::Response;
 use super::facade::WebClientFacade;
 use super::repository_client::HttpClientRepository;
-use crate::app::services::request::entities::{RequestData, METHODS};
+use crate::app::services::request::entities::methods::METHODS;
+use crate::app::services::request::entities::requests::RequestData;
 
 pub type WebClientInstance = Box<dyn WebClientFacade>;
 
@@ -26,12 +27,12 @@ impl WebClientFacade for WebClient {
         } = request;
 
         match request.method {
-            METHODS::GET => self.http_client.call_get(url, headers),
-            METHODS::POST => self.http_client.call_post(url, headers, body),
-            METHODS::PUT => self.http_client.call_put(url, headers, body),
-            METHODS::PATCH => self.http_client.call_patch(url, headers, body),
-            METHODS::HEAD => self.http_client.call_head(url, headers, body),
-            METHODS::DELETE => self.http_client.call_delete(url, headers, body),
+            METHODS::GET => self.http_client.call_get(url.to_string(), headers),
+            METHODS::POST => self.http_client.call_post(url.to_string(), headers, body),
+            METHODS::PUT => self.http_client.call_put(url.to_string(), headers, body),
+            METHODS::PATCH => self.http_client.call_patch(url.to_string(), headers, body),
+            METHODS::HEAD => self.http_client.call_head(url.to_string(), headers, body),
+            METHODS::DELETE => self.http_client.call_delete(url.to_string(), headers, body),
         }
     }
 }
