--- tests/view/snapshots/r#mod__view__input_parser__should_persist_snapshot_to_edit-2.snap.new
@@ -1,14 +1,0 @@
----
-source: tests/view/input_parser.rs
-assertion_line: 93
-expression: result
----
-- SaveRequest:
-    request_name: url.com
-    request_data:
-      url: ~
-      method: ~
-      headers: ~
-      body: "{\"Hello\":\"World\",\"name\":\"Bob\"}"
-    check_exists_before: true
-
